feature: Rework header nav component
By putting more things into the header's nav component,
we can reduce the whitespace between the header's h1 element and
the links, allowing us to also remove the page break element hack
up top.

This has been tested with a tagline and the mugshot. It makes both
appear smaller, which is fine for me.