tech: Add a Runfile
Adding this Runfile in common with some other projects means
that working on this site can happen on other machines, and
isn't dependent on me remembering `deno` commands.

Sadly not POSIX shell, it does need `bash` on account
of the `shift` command.

Commit also amends workflows to use new Runfile, and updates
documentation.